requirements.txt
openai-whisper>=20231117
pyannote.audio>=3.1.1
pydub>=0.25.1
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.1
librosa>=0.10.1
scipy>=1.10.0
numpy>=1.24.0
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.7.0
pathlib>=1.0.1 